fix: split thinking utilities to prevent Node.js code leaking into browser bundle
Extract browser-safe thinking level utilities into thinking-shared.ts
(no Node.js or plugin-runtime dependencies). The Control UI now imports
from thinking-shared.ts instead of thinking.ts, breaking the import chain
that pulled process.env/fs.constants into the Vite browser bundle and
caused a blank-page crash.

Fixes #47998
